restructured files and fixed links
