+++ src/components/common/navBar.vue
