+++ src/components/common/reviewText.vue
@@ -3,7 +3,7 @@
 </template>
 <script lang="ts">
 	import { Vue, Component, Prop, Watch, Lifecycle } from "av-ts";
-	import { LocalisedStrings } from "../util/localisedStrings";
+	import { LocalisedStrings } from "../../util/localisedStrings";
 
 	@Component
 	export default class ReviewText extends Vue {
