+++ src/components/venues/venuePage.vue
@@ -75,13 +75,13 @@
 	<script lang="ts">
 	import { Vue, Component, Lifecycle } from "av-ts";
 	import { mdSnackbar } from "vue-material";
-	import ReviewText from "components/reviewText";
-	import Rating from "components/rating";
-	import ErrorPage from "components/errorPage";
-	import { LocalisedStrings } from "../util/localisedStrings";
-	import { TempReviews } from "../util/tempReviews";
+	import ReviewText from "../common/reviewText";
+	import Rating from "./rating";
+	import ErrorPage from "../common/errorPage";
+	import { LocalisedStrings } from "../../util/localisedStrings";
+	import { TempReviews } from "../../util/tempReviews";
 	import { Review, Venue } from "API";
-	import { Request } from "../util/parmyayRequest";
+	import { Request } from "../../util/parmyayRequest";
 
 	@Component({
 		components: {
